+++ packages/components/src/snackbar/types.ts
@@ -6,11 +6,7 @@ import type { MutableRefObject, ReactNode } from 'react';
 /**
  * Internal dependencies
  */
-import type {
-	NoticeProps,
-	NoticeChildren,
-	NoticeAction,
-} from '../notice/types';
+import type { NoticeProps, NoticeAction } from '../notice/types';
 
 type SnackbarOnlyProps = {
 	/**
@@ -56,10 +52,11 @@ export type SnackbarProps = Pick<
 		 *
 		 * @default []
 		 */
-		actions?: Pick<
-			NoticeAction,
-			'label' | 'url' | 'onClick' | 'openInNewTab'
-		>[];
+		actions?: Array<
+			Pick< NoticeAction, 'label' | 'url' | 'onClick' > & {
+				openInNewTab?: boolean;
+			}
+		>;
 	};
 
 export type SnackbarListProps = {
@@ -70,5 +67,5 @@ export type SnackbarListProps = {
 		}
 	>;
 	onRemove: ( id: string ) => void;
-	children?: NoticeChildren | Array< NoticeChildren >;
+	children?: ReactNode;
 };