+++ packages/dataviews/src/hooks/use-form-validity.ts
@@ -14,7 +14,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import normalizeFields from '../field-types';
-import normalizeForm from '../dataform-layouts/normalize-form';
+import normalizeForm from '../components/dataform-layouts/normalize-form';
 import type {
 	Field,
 	FieldValidity,