+++ test/storybook-playwright/storybook/main.js
@@ -5,10 +5,7 @@ const baseConfig = require( '../../../storybook/main' );
 
 const config = {
 	...baseConfig,
-	addons: [
-		'@storybook/addon-toolbars',
-		'@storybook/addon-webpack5-compiler-babel',
-	],
+	addons: [ '@storybook/addon-webpack5-compiler-babel' ],
 	docs: undefined,
 	staticDirs: undefined,
 	stories: [