+++ packages/block-library/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 -   Post Date Block: Improve semantic HTML to add `<time>` inside link element.
 
+## 9.37.0 (2025-12-23)
+
 ## 9.36.0 (2025-11-26)
 
 ## 9.35.0 (2025-11-12)